+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.9.0] - 2020-09-08
+
 ### Added
 - Update schema definition to version 10.0 (#3)
 - Button in Admin Panel to verify structured data in Rich Result Test
@@ -200,7 +202,8 @@ Initial release
 - View helpers for usage in Fluid templates
 
 
-[Unreleased]: https://github.com/brotkrueml/schema/compare/v1.8.0...HEAD
+[Unreleased]: https://github.com/brotkrueml/schema/compare/v1.9.0...HEAD
+[1.9.0]: https://github.com/brotkrueml/schema/compare/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/brotkrueml/schema/compare/v1.7.2...v1.8.0
 [1.7.2]: https://github.com/brotkrueml/schema/compare/v1.7.1...v1.7.2
 [1.7.1]: https://github.com/brotkrueml/schema/compare/v1.7.0...v1.7.1